Updated camera control for two cameras
The new cli has the ability to set a minimum height for aircraft.  This also now allows two cameras to face north and south and aircraft are passed-off to the other camera when they cross the determined latitude.  I've also included the following key presses - 's' saves a screen-shot, 'r' - restarts the camera (use if camera stops moving or doesn't appear to be following.)  Added a basic compass rose to the bottom of the screen.